queueit_automation/requirements.txt
requests>=2.25.1
uuid>=1.30
tensorflow>=2.9.0
pillow>=9.0.0
numpy>=1.19.5 